+++ features/Context/EnterpriseCommandContext.php
@@ -152,7 +152,9 @@ class EnterpriseCommandContext extends CommandContext
                 );
             }
 
-            assertEquals($proposal->getChanges(), $expectedResult);
+            $changes = $this->sanitizeDraftFileChanges($proposal->getChanges());
+
+            assertEquals($changes, $expectedResult);
         }
     }
 
@@ -326,6 +328,30 @@ class EnterpriseCommandContext extends CommandContext
                 $diff
             );
         }
+    }
+
+    /**
+     * @param array $changes
+     *
+     * @return array
+     */
+    protected function sanitizeDraftFileChanges(array $changes)
+    {
+        foreach ($changes['values'] as $attributeCode => $change) {
+            foreach ($change as $changeKey => $data) {
+                $data = $data['data'];
+                if (isset($data['filePath']) && isset($data['originalFilename'])) {
+                    $filenameLength = strlen($data['originalFilename']);
+                    $data['filePath'] = substr($data['filePath'], -$filenameLength);
+                }
+                if (isset($data['hash'])) {
+                    unset($data['hash']);
+                }
+                $changes['values'][$attributeCode][$changeKey]['data'] = $data;
+            }
+        }
+
+        return $changes;
     }
 
     /**