+++ spec/PimEnterprise/Bundle/WorkflowBundle/Presenter/FilePresenterSpec.php
@@ -2,8 +2,10 @@
 
 namespace spec\PimEnterprise\Bundle\WorkflowBundle\Presenter;
 
+use Akeneo\Component\FileStorage\Model\FileInterface;
 use PhpSpec\ObjectBehavior;
-use Pim\Bundle\CatalogBundle\Model;
+use Pim\Bundle\CatalogBundle\Model\AttributeInterface;
+use Pim\Bundle\CatalogBundle\Model\ProductValueInterface;
 use Symfony\Component\Routing\Generator\UrlGeneratorInterface;
 
 class FilePresenterSpec extends ObjectBehavior
@@ -19,8 +21,8 @@ class FilePresenterSpec extends ObjectBehavior
     }
 
     function it_supports_file(
-        Model\ProductValueInterface $value,
-        Model\AttributeInterface $attribute
+        ProductValueInterface $value,
+        AttributeInterface $attribute
     ) {
         $attribute->getAttributeType()->willReturn('pim_catalog_file');
         $value->getAttribute()->willReturn($attribute);
@@ -30,33 +32,26 @@ class FilePresenterSpec extends ObjectBehavior
 
     function it_presents_files_in_a_list(
         $generator,
-        Model\ProductValueInterface $value,
-        Model\ProductMedia $media
+        ProductValueInterface $value,
+        FileInterface $media
     ) {
-        $file2 = new\SplFileInfo(__FILE__);
-
         $value->getMedia()->willReturn($media);
-        $media->getFilename()->willReturn(__FILE__);
+        $media->getKey()->willReturn('key/of/the/media.pdf');
+        $media->getHash()->willReturn('54qsfda8e7r54f');
         $media->getOriginalFilename()->willReturn('original_foo.pdf');
 
         $generator
-            ->generate('pim_enrich_media_show', ['filename' => __FILE__], true)
-            ->willReturn(__FILE__);
+            ->generate('pim_enrich_media_show', ['filename' => urlencode('key/of/the/media.pdf')])
+            ->willReturn('url/of/the/media.pdf');
         $generator
-            ->generate('pim_enrich_media_show', ['filename' => __FILE__])
-            ->willReturn(__FILE__);
-        $generator
-            ->generate('pim_enrich_media_show', ['filename' => $file2->getFilename()], true)
-            ->willReturn($file2->getPath());
-        $generator
-            ->generate('pim_enrich_media_show', ['filename' => $file2->getFilename()])
-            ->willReturn($file2->getPath());
+            ->generate('pim_enrich_media_show', ['filename' => urlencode('key/of/the/change.txt')])
+            ->willReturn('url/of/the/change.txt');
 
         $change = [
             'data' => [
-                'filename' => $file2->getFilename(),
-                'filePath' => $file2->getPath(),
-                'originalFilename' => 'change_foo.pdf',
+                'hash' => '98az7er654ert4s',
+                'filePath' => 'key/of/the/change.txt',
+                'originalFilename' => 'the new change.txt',
             ]
         ];
 
@@ -68,33 +63,29 @@ class FilePresenterSpec extends ObjectBehavior
                 '</li>' .
                 '<li class="changed file">' .
                     '<i class="icon-file"></i>' .
-                    '<a target="_blank" class="no-hash" href="%s">change_foo.pdf</a>' .
+                    '<a target="_blank" class="no-hash" href="%s">the new change.txt</a>' .
                 '</li>' .
             '</ul>',
-            __FILE__,
-            $file2->getPath()
+            'url/of/the/media.pdf',
+            'url/of/the/change.txt'
         ));
     }
 
-    function it_only_presents_new_file_if_value_does_not_have_a_media_yet(
+    function it_only_presents_new_file_if_value_does_not_has_a_media_yet(
         $generator,
-        Model\ProductValueInterface $value
+        ProductValueInterface $value
     ) {
         $value->getMedia()->willReturn(null);
 
-        $file = new \SplFileInfo(__FILE__);
         $generator
-            ->generate('pim_enrich_media_show', ['filename' => $file->getFilename()], true)
-            ->willReturn($file->getPath());
-        $generator
-            ->generate('pim_enrich_media_show', ['filename' => $file->getFilename()])
-            ->willReturn('/media/uploaded_foo.pdf');
+            ->generate('pim_enrich_media_show', ['filename' => urlencode('key/of/the/change.txt')])
+            ->willReturn('url/of/the/change.txt');
 
         $change = [
             'data' => [
-                'filename' => $file->getFilename(),
-                'filePath' => $file->getPath(),
-                'originalFilename' => 'foo.pdf',
+                'hash'             => '98az7er654ert4s',
+                'filePath'         => 'key/of/the/change.txt',
+                'originalFilename' => 'the new change.txt',
             ]
         ];
 
@@ -102,7 +93,7 @@ class FilePresenterSpec extends ObjectBehavior
             '<ul class="diff">' .
                 '<li class="changed file">' .
                     '<i class="icon-file"></i>' .
-                    '<a target="_blank" class="no-hash" href="/media/uploaded_foo.pdf">foo.pdf</a>' .
+                    '<a target="_blank" class="no-hash" href="url/of/the/change.txt">the new change.txt</a>' .
                 '</li>' .
             '</ul>'
         );
@@ -110,26 +101,23 @@ class FilePresenterSpec extends ObjectBehavior
 
     function it_only_presents_old_file_if_a_new_one_is_not_provided(
         $generator,
-        Model\ProductValueInterface $value,
-        Model\ProductMedia $media
+        ProductValueInterface $value,
+        FileInterface $media
     ) {
-        $file = new \SplFileInfo(__FILE__);
         $value->getMedia()->willReturn($media);
-        $media->getFilename()->willReturn($file->getFilename());
-        $media->getOriginalFilename()->willReturn('bar.pdf');
+        $media->getKey()->willReturn('key/of/the/media.pdf');
+        $media->getHash()->willReturn('54qsfda8e7r54f');
+        $media->getOriginalFilename()->willReturn('original_foo.pdf');
 
         $generator
-            ->generate('pim_enrich_media_show', ['filename' => $file->getFilename()], true)
-            ->willReturn($file->getPath());
-        $generator
-            ->generate('pim_enrich_media_show', ['filename' => $file->getFilename()])
-            ->willReturn('/media/uploaded_bar.pdf');
+            ->generate('pim_enrich_media_show', ['filename' => urlencode('key/of/the/media.pdf')])
+            ->willReturn('url/of/the/media.pdf');
 
         $this->present($value, ['data' => []])->shouldReturn(
             '<ul class="diff">' .
                 '<li class="base file">' .
                     '<i class="icon-file"></i>' .
-                    '<a target="_blank" class="no-hash" href="/media/uploaded_bar.pdf">bar.pdf</a>' .
+                    '<a target="_blank" class="no-hash" href="url/of/the/media.pdf">original_foo.pdf</a>' .
                 '</li>' .
             '</ul>'
         );